// src/api/generated/models/requirementRevisionResponseDto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Requirements Backend API
 * HTTP API for the Requirements Management app.
 * OpenAPI spec version: 0.0.1
 */
import type { RequirementResponseDto } from './requirementResponseDto';

export type RequirementRevisionResponseDto = RequirementResponseDto & {
  requirementId: string;
  /** @minimum 1 */
  revisionNumber: number;
  requirementCreatedAt: string;
  requirementUpdatedAt: string;
};